update again validation scripts

+++ libminipic/validation/beam_heavy.py
@@ -296,7 +296,7 @@ def validate(evaluate=True, threshold=THRESHOLD):
                 minipic_diag.read_1d_diag("diags/" + file)
             )
 
-            new_data.append(np.sum(data * x_data))
+            new_data.append(float(np.sum(data * x_data)))
 
         print("    - For species {}: {}".format(ispecies, new_data))
